[wip] Move Fuchsia to Embedder API
This allows removing *all* Fuchsia-specific code from flow/, and
removing all non-embedder dependencies from the flutter_runner.

It also decouples the compositor's scene graph from Flutter's scene
graph, massively reducing render workload and memory allocations.

As a bonus, we get to delete thousands of lines of old code that
was only around for Fuchsia :)

+++ flow/layers/performance_overlay_layer.h
@@ -15,10 +15,10 @@ class SkTextBlob;
 
 namespace flutter {
 
-const int kDisplayRasterizerStatistics = 1 << 0;
-const int kVisualizeRasterizerStatistics = 1 << 1;
-const int kDisplayEngineStatistics = 1 << 2;
-const int kVisualizeEngineStatistics = 1 << 3;
+const int64_t kDisplayRasterizerStatistics = 1 << 0;
+const int64_t kVisualizeRasterizerStatistics = 1 << 1;
+const int64_t kDisplayEngineStatistics = 1 << 2;
+const int64_t kVisualizeEngineStatistics = 1 << 3;
 
 class PerformanceOverlayLayer : public Layer {
  public:
@@ -26,13 +26,14 @@ class PerformanceOverlayLayer : public Layer {
                                               const std::string& label_prefix,
                                               const std::string& font_path);
 
-  explicit PerformanceOverlayLayer(uint64_t options,
+  explicit PerformanceOverlayLayer(int64_t options,
                                    const char* font_path = nullptr);
 
+  // |Layer|
   void Paint(PaintContext& context) const override;
 
  private:
-  int options_;
+  int64_t options_;
   std::string font_path_;
 
   FML_DISALLOW_COPY_AND_ASSIGN(PerformanceOverlayLayer);
